+++ include/bitcoin/blockchain/interface/safe_chain.hpp
@@ -146,7 +146,7 @@ public:
     //-------------------------------------------------------------------------
 
     virtual void fetch_template(inventory_fetch_handler handler) const = 0;
-    virtual void fetch_mempool(size_t count_limit, size_t minimum_fee,
+    virtual void fetch_mempool(size_t count_limit, uint64_t minimum_fee,
         inventory_fetch_handler handler) const = 0;
 
     // Filters.
